Refactor and add requirements
